+++ CHANGE_LOG.md
@@ -29,6 +29,7 @@ information about obtaining the commercial api credentials.
 * Fix error message when trying to add a subreddit without an existing subreddit list.
 * Fix imgur url formatting that happened in some circumstances
 * Fix occasional crashing due to posts removed by reddit
+* Fix crashing when trying to download from private subreddits